Fixes test.

+++ MagibooksKata/src/main/java/DiscountCalculator.java
@@ -1,3 +1,3 @@
 public interface DiscountCalculator {
-    double getDiscountAmountPounds();
+    float getDiscountAmountPounds();
 }
